Make image copying modular
This patch makes the process of copying the image modular.

Until now snf-image expected the target disk to be a local
block device. But since Ganeti supports userspace access to disks
(e.g. via the RBD and the ExtStorage disk template) we should
not take that for granted. Until now the image copying works
because the target disk is a block device. In case of RADOS
without having the volume locally mapped this would not work.
Additionally in case of Archipelago if we choose to go only
with userspace support (without using blktap to create a local
block device) this would not work either.

This patch abstracts the way snf-image handles source images
and destination disks and makes the whole process modular.

snf-image obtains the source image from the IMG_ID environment
variable and the destination disk from DISK_0_PATH if it is a
valid block device otherwise from DISK_0_URI.

It separates source backends from destination backends and
supports execution of external scripts tailored to each
backend type.

snf-image currently knows the following source backends (for images):

  - null that practically skips data copy
  - file for local raw images
  - network for images that can be downloaded via curl
  - pithos for images stored in pithos and downloaded with pithcat

and the following destination backends (for disks):

  - file for disks that are either local files or local block devices

This patch introduces the following helper functions:

  - get_src_backend_type() returns the backend type of the source
    image based on IMG_ID

  - get_dst_backend_type() returns the backend type of the
    destination disk based on DISK_0_PATH and DISK_0_URI.
    If we have a valid DISK_0_PATH then the default file
    backend is chosen.

So the copy procedures boils down to:

  SRC_BACKEND_TYPE=$(get_src_backend_type)
  DST_BACKEND_TYPE=$(get_dst_backend_type)

  src_script=backends/src/${SRC_BACKEND_TYPE}
  dst_script=backends/dst/${DST_BACKEND_TYPE}

  image_size=$($src_script -s $IMG_ID $IMG_FORMAT)
  monitor="./copy-monitor.py -o $MONITOR_FD -r $image_size"
  $src_script $IMG_ID $IMG_FORMAT | $monitor dd | $dst_script $disk0 $IMG_FORMAT

A custom backend should define two scripts named after the
backend type:

 - A source script that will take $IMG_ID and $IMG_FORMAT and if
   the '-s' option gets passed it returns the image size otherwise
   it dumps the image data to stdout

 - A destination script that will take $disk0 and $IMG_FORMAT and will
   read data from stdin and dump them to $disk0 taking into account the
   different image types (diskdump, extdump or ntfsdump).

snf-image will use a custom source backend (for images) if $IMG_ID
is of the following format: <backend type>:<some custom identifier>.
The custom script is expected to be the
`/usr/share/ganeti/os/snf-image/backends/src/<backend type>'

Accordingly, it will use a custom destination backend (for the
target disk) if the exported by Ganeti $DISK_0_PATH is missing and
the $DISK_0_URI is of the same format as above. The current
implementation in Ganeti sets $DISK_0_URI to the value that
eventually will be passed to the -drive option in the QEMU command
line (e.g.
archipelago:0e56cebe-40bc-46d9-b03a-76acb428ba15.ext.disk0).
The custom script is expected to be the
`/usr/share/ganeti/os/snf-image/backends/dst/<backend type>'.

This patch moves handling of file system dumps to the destination
script of the file backend.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>